refactor(driver): extract tx_journal + l1_anchor impl methods to driver/journal.rs (refactor 2.1f)
First impl-block extraction of Phase 2.1. Moves the six persistence
methods (tx journal + L1-confirmed anchor) out of `impl Driver` in
mod.rs into a second `impl Driver` block in a sibling module.

- `crates/based-rollup/src/driver/journal.rs` (new, 250 lines):
  - Full `impl<P, Pool> Driver<P, Pool> where …` block with the same
    where-clause as mod.rs (reth provider + tx pool bounds).
  - Methods migrated: `journal_block_transactions`, `save_tx_journal`,
    `load_tx_journal`, `prune_tx_journal`, `save_l1_confirmed_anchor`,
    `load_l1_confirmed_anchor`.
  - All methods are `pub(super) fn` so the main `impl Driver` block
    in mod.rs can still call them via `self.save_tx_journal()` etc.
    The previously-private fields (`tx_journal`, `l2_provider`,
    `l1_confirmed_anchor`, `pending_reinjection`, `l2_head_number`,
    `config`) remain accessible because `journal` is a child module
    of `driver`, and Rust's privacy rules grant child modules access
    to the parent's private items.
  - Imports pulled in: `super::Driver`, `super::types::{L1ConfirmedAnchor,
    TX_JOURNAL_STAGE_ID, TxJournalEntry}`,
    `crate::derivation::{L1_CONFIRMED_L1_STAGE_ID, L1_CONFIRMED_L2_STAGE_ID}`,
    `reth_provider::*` bounds, `reth_stages_types::StageCheckpoint`,
    `reth_primitives_traits::SignerRecoverable` (for `recover_signer`),
    `tracing::{debug, info, warn}`.

- `crates/based-rollup/src/driver/mod.rs` (-182 lines → 4387 total):
  - `mod journal;` declaration.
  - All six method bodies removed from the main `impl Driver` block.
  - Removed now-unused imports: `L1_CONFIRMED_L1_STAGE_ID`,
    `L1_CONFIRMED_L2_STAGE_ID`, `DBProvider`, `StageCheckpoint`,
    `TX_JOURNAL_STAGE_ID` (the latter is still imported in
    journal.rs via `super::types::`).

Cumulative progress:
- mod.rs: 5509 → 4387 lines (-1122 / -20.4%)
- 5 sibling modules totaling 1251 lines of well-bounded code.

`cargo check`, `cargo clippy --all-features -- -D warnings`, and
`cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
change — every `self.save_tx_journal()` call site resolves to the
same method body, just in a different file.

Next: 2.1g extracts the rewind methods (`set_rewind_target`,
`rewind_to_re_derive`, `clear_internal_state`, `rewind_l2_chain`) to
`driver/rewind.rs`.
